Added handling of "zombie" squashed commits, add text indicator of squashed, added null check before deletion of squash commit

+++ packages/api/src/gitlab/merge-request/merge-request.service.ts
@@ -198,19 +198,17 @@ export class MergeRequestService extends BaseService<
     repository: Repository,
     mergeRequest: MergeRequestEntity,
   ) {
-    let commits = await this.fetchCommitsFromGitlab(
+    const commits = await this.fetchCommitsFromGitlab(
       token,
       repository,
       mergeRequest.resource,
     );
     if (mergeRequest.resource.squash) {
-      // commits = commits.map((commit) => {
-      //   commit.resource.extensions.squashed = true;
-      //   return commit;
-      // })
       await this.commitService.syncForRepositoryPage(token, repository, commits, true);
       let squashCommit = await this.commitService.findByGitlabId(repository, mergeRequest.resource.squash_commit_sha);
-      await this.commitService.deleteCommitEntity(squashCommit);
+      if (squashCommit) {
+        await this.commitService.deleteCommitEntity(squashCommit);
+      }
     }
     mergeRequest.commits = await Promise.all(
       commits.map((commit) =>
